src/test/java/com/foodrecord/core/db/routing/AdaptiveDataSourceRouterTest.java
package com.foodrecord.core.db.routing;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import static org.mockito.Mockito.*;
import static org.assertj.core.api.Assertions.*;

@ExtendWith(MockitoExtension.class)
class AdaptiveDataSourceRouterTest {

//    @Mock
//    private PerformanceMonitor performanceMonitor;
//
//    @Mock
//    private LoadPredictor loadPredictor;
//
//    @Mock
//    private QueryPatternAnalyzer queryAnalyzer;
//
//    @InjectMocks
//    private AdaptiveDataSourceRouter router;
//
//    @Test
//    void shouldSelectOptimalDataSourceForRead() {
//        // Given
//        TransactionContext context = mock(TransactionContext.class);
//        when(context.getSql()).thenReturn("SELECT * FROM users");
//
//        QueryCharacteristics queryChars = mock(QueryCharacteristics.class);
//        when(queryAnalyzer.analyze(anyString())).thenReturn(queryChars);
//
//        LoadImpact impact = mock(LoadImpact.class);
//        when(loadPredictor.predictImpact(any())).thenReturn(impact);
//
//        Map<DataSource, PerformanceMetrics> metrics = new HashMap<>();
//        when(performanceMonitor.getCurrentMetrics()).thenReturn(metrics);
//
//        // When
//        DataSource selectedDataSource = router.determineReadDataSource(context);
//
//        // Then
//        assertThat(selectedDataSource).isNotNull();
//        verify(queryAnalyzer).analyze(anyString());
//        verify(loadPredictor).predictImpact(any());
//        verify(performanceMonitor).getCurrentMetrics();
//    }
//
//    @Test
//    void shouldHandleHighLoadScenario() {
//        // Given
//        TransactionContext context = mock(TransactionContext.class);
//        LoadImpact impact = mock(LoadImpact.class);
//        when(impact.getCpuImpact()).thenReturn(0.8);
//        when(loadPredictor.predictImpact(any())).thenReturn(impact);
//
//        // When
//        DataSource selectedDataSource = router.determineReadDataSource(context);
//
//        // Then
//        assertThat(selectedDataSource).isNotNull();
//        verify(loadPredictor).predictImpact(any());
//    }
} 